Fix type error
