+++ packages/editor/src/state/IFrameRPC.ts
@@ -1,0 +1,14 @@
+export interface IRootToEditorRPCHandler {
+  sync(data: Uint8Array): Promise<void>;
+  init(data: Uint8Array): Promise<void>;
+}
+
+export interface IEditorToRootRPCHandler {
+  ready(): Promise<void>;
+  update(data: Uint8Array): Promise<void>;
+  uploadImage(
+    hash: string,
+    contentType: string,
+    data: Uint8Array
+  ): Promise<string>;
+}
